requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
psutil==7.0.0
PyYAML==6.0.2
Jinja2==3.1.6
